switching from simp to pure_wp.  Main modifications:
`pure` is now based on `pure_wp` instead of simp, the proof mode lemmas are
adapted (most importantly in `simp_eval.v`, which is now not split into `simp`
and `pure` lemmas)

Temporary `total.v` file shadowing `simplification`'s `total` (and associated
lemmas) with `pure_wp`, so as to be used transparently elsewhere (but really
this interface should be removed later, once decided which names are best).

Note that the reciprocal of the bind rule does not need the `stack` relation,
and so `simplification.v` can probably be simplified quite a lot.

Some clean up, moved "`pure_wp` implies ewp" to `basic_rules.v`.

Admitted lemma in splay.v (quaternary version of a sequentialization lemma)